fix(lib): rework positioning of selection box
This commit reworks the positioning of the selection box completely and removes
the need for a fixed position. The select-container is now relative and the selection
box is positioned absolute. With that being changed, the library now correctly renders
the selection box even while scrolling. Instead of using pageX and pageY the coordinates
are now calculated using clientX and clientY, which represent the mouse position relative to the browser's visible viewport and not to the html document. This is also the W3C recommendation.

Addresses #8