feat(lib): add classes for selection modes
This feature introduces two classes - dts-adding and dts-removing - which are
added to the select box element for more flexibility in styling. In addition,
the color of the select box when in removing mode can now be styled with a
separate variable '$select-box-removing-color'. If the color should be the same
for both adding and removing, the '$dts-primary' variable can be used.

Fixes #52