fix(lib): prevent key events from being fired repeatedly
This commit fixes the issue with keyboard events being fired repeatedly on Windows when a key is being pressed and hold pressed. This was not the case on Mac and there was only one event triggered per keydown event, even if the key was hold pressed. The issue is resolved with a custom operator that only emits the keyboard event if the `keyCode` is different from the last one.

Closes #61