feat(lib): add selection outputs
This commit adds two outputs to the select-container - `selectionStarted` and `selectionEnded`. This helps with better and more efficient  processing of selected items. There is `select` but this output is emitted with each and every selected item. For large collections this is quite inefficient. With the new outputs this becomes much better.

Closes #64