fix(lib): correctly select items when selectOnDrag is false
This commit fixes the issue of items not being selected when `selectOnDrag` is set to `false`. The problem was that when `selectOnDrag` is disabled, there were two functions that took care of selecting an item - on mousedown and mouseup. However, the one for mouseup was supposed to be used only when the user is dragging and the select box is visible. If the select box was  not visible, the handler for mouseup caused the item to be deselected again. The issue is fixed by only triggering the mouseup handler when the user is dragging.

Closes #67